% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fritools2-package.R
\docType{package}
\name{fritools2-package}
\alias{fritools2-package}
\title{Utilities for the Forest Research Institute of the State
Baden-Wuerttemberg}
\description{
Miscellaneous utilities, tools and helper functions.
}
\details{
You will find the details in\cr
\code{vignette("Not_an_Introduction_to_fritools", package = "fritools2")}.
}
\keyword{package}
