% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_package.R
\name{rename_package}
\alias{rename_package}
\title{Change This Package's Name}
\usage{
rename_package(path, to)
}
\arguments{
\item{path}{The new package name.}
}
\value{
Run for its side effect.
}
\description{
Does not change the path, we just rename the package, not the repo
and we must not change git remotes!
}
\keyword{internal}
